Se modifica el campo valor en la migracion de la trm, ya que habia un bug de redondeado

+++ database/migrations/2019_08_29_145814_create_trm_table.php
@@ -15,7 +15,7 @@ class CreateTrmTable extends Migration
     {
         Schema::create('trm', function (Blueprint $table) {
             $table->bigIncrements('id');
-						$table->float('value');
+						$table->double('value');
 						$table->boolean('status')->default(1);
             $table->timestamps();
         });
